==> SIOSE AR/01 - merging and uses extraction.ipynb <==
### Extracción de capas de interés de SIOSE
#### (Solar, Cultivos y Cultivos arbóreos)

##### Instalar paquetes no nativos

!pip install geopandas pyproj

!pip install fiona

##### Importar paquetes

import os
import glob
import geopandas as gpd
from pyproj import CRS
import zipfile
import fiona
import pandas as pd
import pickle
import pandas as pd

##### Listar todos los archivos de SIOSE, formato .gpkg para cada provincia

# Directorio raíz
directorio_raiz = "spatial_data/"

# Lista para almacenar los archivos .gpkg
gpkg_files = []

# Recorrer recursivamente el directorio raíz y sus subdirectorios
for directorio_actual, _, archivos in os.walk(directorio_raiz):
    # Filtrar los archivos .gpkg y agregarlos a la lista
    gpkg_files += [os.path.join(directorio_actual, archivo) for archivo in archivos if archivo.endswith(".gpkg")]

# Imprimir la lista de archivos .gpkg
print(gpkg_files)

##### Extracción polígonos solares

# Función para extraer la capa USOS y el uso 'Solar'
def extract_USOS(file_path):
    # Listar todas las capas
    layers = fiona.listlayers(file_path)
    # Seleccionar el nombre de capa que termine por '_T_USOS'
    layer_name = [layer for layer in layers if layer.endswith('_T_USOS')][0]
    # Leer la capa
    layer = gpd.read_file(file_path, layer=layer_name)
    # Extraer uso 'Solar'
    layer = layer[layer['ID_USO_MAX'].isin([2442])]
    return layer

# Extraer capas 
solar_polygons = [extract_USOS(file) for file in gpkg_files]

# Proyectar a EPSG:32630
solar_polygons = [capa.to_crs(32630) for capa in solar_polygons]

# Iterar sobre las capas de la lista para ver cuáles tienen más polígonos
for i, geo_database in enumerate(solar_polygons, start=1):
    num_polygons = len(geo_database)  # Obtener la cantidad de polígonos en la geodatabase actual
    print(f"La geodatabase {i} contiene {num_polygons} polígonos.")

# La que más tiene es Granada, pero nos quedamos con Badajoz que es la segunda que más tiene
# porque los polígonos de Granada creo que de media son mucho más pequeños

# Guardar variable de Python como archivo (pickle package)

# Ruta del archivo
ruta_archivo = "spatial_data/solar_polygons.pkl"

# Guardar la variable en el archivo
with open(ruta_archivo, "wb") as archivo:
    pickle.dump(solar_polygons, archivo)

print("Variable guardada correctamente en", ruta_archivo)

# Así podemos cargar la variable en un nuevo entorno


# En caso de necesitar rehacer proceso podemos volver a cargar la variable
# Abre el archivo en modo lectura binaria
with open('spatial_data/solar_polygons.pkl', 'rb') as file:
    solar_polygons = pickle.load(file)

# Seleccionar Badajoz y Ciudad Real y unir
sub_solar = pd.concat([solar_polygons[4], solar_polygons[11]], ignore_index=True)
print(sub_solar)

# Guardar
sub_solar.to_file("spatial_data/final_polys/badajoz_creal_solar.shp")

##### Extracción polígonos invernaderos

# Continuamos sólo con Badajoz para facilitar el proceso
capas_badajoz=fiona.listlayers("spatial_data/06_BADAJOZ.gpkg")
capas_badajoz

# La capa que nos interesa es la de POLIGONOS, donde se encuentran los usos agrícolas detallados según SIGPAC
badajoz_POLIGONOS = gpd.read_file("spatial_data/06_BADAJOZ.gpkg", layer=capas_badajoz[1])

# Nuestros polígonos de interés son los que en la columna de SIGPAC tienen la categoría 'IV'
# Extraer polígonos, proyectar y guardar
badajoz_greenhouse = badajoz_POLIGONOS[badajoz_POLIGONOS['USO_SIGPAC'].isin(['IV'])].to_crs(25830)
badajoz_greenhouse.to_file("spatial_data/final_polys/badajoz_greenhouse.shp")

##### Extracción polígonos cultivos arbóreos

badajoz_POLIGONOS['USO_SIGPAC'].unique()


badajoz_treecrops = badajoz_POLIGONOS[badajoz_POLIGONOS['USO_SIGPAC'].isin(['CF', 'CI', 'CS', 'CV', 'FF',
                                                                          'FL', 'FS', 'FV', 'FY', 'OC',
                                                                          'OF', 'OV', 'VF', 'VI', 'VO'])].to_crs(25830)
badajoz_treecrops.to_file("spatial_data/final_polys/badajoz_treecrops.shp")

##### Extracción polígonos urbanos

# Nuestra capa de interés es COMBINADA
badajoz_COMBINADA = gpd.read_file("spatial_data/06_BADAJOZ.gpkg", layer=capas_badajoz[2])

# Seleccionamos los polígonos cuya 'COBERTURA_DESC' sea 'Edificación'
badajoz_urban = badajoz_COMBINADA[badajoz_COMBINADA['COBERTURA_DESC'].isin(['Edificación'])].to_crs(25830)
badajoz_urban.to_file("spatial_data/final_polys/badajoz_urban.shp")

##### Extracción todos los demás polígonos

# esto no hace falta hacerlo porque realmente es asignar la categoría 'otros' a todos los píxeles
# que no correspondan a ninguna de las anteriores así que ya está!!
